refactor[db]: Schema and entity configurations
Refactored the database and entity configurations to build off a general
base entity configuration. It both simplifies the code and separates the
individual entity configurations, so it is easier to change later on.

+++ service/DTOs/StockDto.cs
@@ -2,7 +2,7 @@ namespace Stoxolio.Service.DTOs;
 
 public class StockDto
 {
-    public int Id { get; init; }
+    public long Id { get; init; }
     public string Name { get; init; } = string.Empty;
     public string Ticker { get; init; } = string.Empty;
     public string Exchange { get; init; } = string.Empty;
@@ -10,9 +10,9 @@ public class StockDto
     public int Shares { get; init; }
     public decimal Price { get; init; }
     public bool Invest { get; init; }
-    public int CategoryId { get; init; }
+    public long CategoryId { get; init; }
     public decimal PrevPrice { get; init; }
-    
+
     // Computed fields
     public decimal Value { get; init; }
     public decimal PriceChange { get; init; }
